refactor: 패키지명 수정
TODO: Stream으로 LottoComparator 관리하기, SetQuantity 사용하지 않기.
