+++ src/main/java/Application.java
@@ -1,4 +1,4 @@
-import lottocontroller.LottoController;
+import lotto_controller.LottoController;
 
 public class Application {
     public static void main(String[] args) {
